refactor: :recycle: update folder structure and move layout pages to modules
