+++ app/[locale]/page.tsx
@@ -1,23 +1,5 @@
-import { Box, Container } from '@mantine/core'
-
-import ContentGrid from '@/components/ContentGrid'
-import Landing from '@/components/Landing'
-
-import classes from './index.module.css'
+import Homepage from '@/modules/Homepage'
 
 export default function Home() {
-  return (
-    <Container
-      fluid
-      p="0"
-    >
-      <Box
-        component="main"
-        className={classes.main}
-      >
-        <Landing />
-        <ContentGrid />
-      </Box>
-    </Container>
-  )
+  return <Homepage />
 }
