--- app/[locale]/posts/[slug]/index.module.css
@@ -1,3 +1,0 @@
-.container {
-  background-color: var(--mantine-color-black);
-}
