+++ modules/Homepage/Landing/index.module.css
@@ -1,5 +1,3 @@
-/* relative flex h-full items-center justify-center bg-black */
-
 .wrapper {
   position: relative;
   display: flex;
