+++ modules/PostDetail/index.module.css
@@ -1,3 +1,4 @@
+.wrapper,
 .container {
   background-color: var(--mantine-color-black);
 }
